feat: get physical dimensions for linux

+++ src/lib.rs
@@ -48,10 +48,10 @@ pub struct DisplayInfo {
     pub width: u32,
     /// The display pixel height.
     pub height: u32,
-    /// The width of a display in millimeters.
-    pub width_mm: f32,
-    /// The height of a display in millimeters.
-    pub height_mm: f32,
+    /// The width of a display in millimeters. This value may be 0.
+    pub width_mm: i32,
+    /// The height of a display in millimeters. This value may be 0.
+    pub height_mm: i32,
     /// Can be 0, 90, 180, 270, represents screen rotation in clock-wise degrees.
     pub rotation: f32,
     /// Output device's pixel scale factor.
